+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- update: Firebase and other dependencies.
+
 
 ## [4.11.0] - 2018-06-20
 - feature: Add importHtml function to <Editor />.
 
 ## [4.10.0] - 2018-06-20
-- Updating prosemirror and react dependencies.
+- update: Prosemirror and react dependencies.
 
 ## [4.9.1] - 2018-06-08
 - feature: Allow images to link to their src.